ai/pyproject.toml
[project]
name = "ai-worker-python"
version = "0.1.0"
description = "Senior 2026 AI Microservice"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.19.0",
    "google-genai>=0.3.0",
    "ollama>=0.1.0",
    "httpx>=0.27.0",
    "opentelemetry-api>=1.23.0",
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "mcp>=0.1.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "vcrpy>=6.0.1",
    "pytest-recording>=0.13.1",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "types-requests>=2.31.0",
    "types-protobuf>=4.24.0",
]

[tool.mypy]
strict = true
ignore_missing_imports = true
exclude = ["generated/"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "C4", "ARG", "PTH"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
